requirements.txt
Flask
pymongo
python-dotenv
werkzeug
bcrypt
